third_party/blink/web_tests/external/wpt/css/css-images/image-set/image-set-parsing.html

<!DOCTYPE html>
<title>Image set parsing</title>
<link rel="author" title="Noam Rosenthal" href="mailto:noam.j.rosenthal@gmail.com">
<link rel="help" href="https://drafts.csswg.org/css-images-4/#image-set-notation">
<meta name="assert" content="General image-set parsing  follows CSS Images 4 rules.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/image-set-parsing-tests.js"></script>
<body>
<script>
function check_image_set(tst) {
  var div = document.createElement('div');
  div.setAttribute("style", `${tst.property}: image-set(${tst.imageSet})`)

  var inline_style = div.style.getPropertyValue(tst.property);
  assert_equals(inline_style.startsWith('image-set'), tst.valid);

  document.body.appendChild(div);
  var computed_style = getComputedStyle(div).getPropertyValue(tst.property);
  assert_equals(computed_style.startsWith('image-set'), tst.valid);

  div.remove();
}

kImageSetTests.forEach(tst => {
  test(() => {
    check_image_set(tst);
  }, `${tst.property}: image-set(${tst.imageSet}) ${tst.valid ? "[ parsable ]" : "[ unparsable ]"}`)
});
</script>
